requirements.txt
requests==2.32.0
pytest==7.4.0
allure-pytest==2.13.2
pytest-rerunfailures==12.0
allure-python-commons~=2.13.2
selenium~=4.29.0